Fix DELETE endpoint testing
